requirements.txt
anyio==3.5.0
apache-beam==2.37.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bertopic==0.9.4
bleach==4.1.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.28
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docopt==0.6.2
elastic-transport==8.1.1
elasticsearch==7.7.1
entrypoints==0.4
executing==0.8.3
faiss-gpu==1.7.2
fastapi==0.75.1
fastavro==1.4.10
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.1
Flask-Compress==1.11
fonttools==4.32.0
grpcio==1.44.0
h11==0.13.0
hdbscan==0.8.28
hdfs==2.7.0
httplib2==0.19.1
huggingface-hub==0.5.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
kiwisolver==1.4.2
llvmlite==0.38.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
nlp==0.4.0
nltk==3.7
notebook==6.4.10
notebook-shim==0.1.0
numba==0.55.1
numpy==1.21.5
oauth2client==4.1.3
orjson==3.6.7
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.7.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
Pygments==2.11.2
pymongo==3.12.3
pynndescent==0.5.6
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
rank-bm25==0.2.2
regex==2022.3.15
requests==2.27.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
starlette==0.17.1
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.8.2+cu111
torchaudio==0.8.2
torchvision==0.9.2+cu111
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.17.0
typing-extensions==4.1.1
umap-learn==0.5.2
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
xxhash==3.0.0
zipp==3.8.0
